Bug 1021828 - Project editor: Show prompt before deleting file. r=paul

+++ browser/devtools/projecteditor/lib/plugins/delete/delete.js
@@ -6,10 +6,12 @@
 
 const { Class } = require("sdk/core/heritage");
 const { registerPlugin, Plugin } = require("projecteditor/plugins/core");
+const { confirm } = require("projecteditor/helpers/prompts");
 const { getLocalizedString } = require("projecteditor/helpers/l10n");
 
 var DeletePlugin = Class({
   extends: Plugin,
+  shouldConfirm: true,
 
   init: function(host) {
     this.host.addCommand({
@@ -22,14 +24,28 @@ var DeletePlugin = Class({
     });
   },
 
+  confirmDelete: function(resource) {
+    let deletePromptMessage = resource.isDir ?
+      getLocalizedString("projecteditor.deleteFolderPromptMessage") :
+      getLocalizedString("projecteditor.deleteFilePromptMessage");
+    return !this.shouldConfirm || confirm(
+      getLocalizedString("projecteditor.deletePromptTitle"),
+      deletePromptMessage
+    );
+  },
+
   onCommand: function(cmd) {
     if (cmd === "cmd-delete") {
       let tree = this.host.projectTree;
       let resource = tree.getSelectedResource();
-      let parent = resource.parent;
+
+      if (!this.confirmDelete(resource)) {
+        return;
+      }
+
       resource.delete().then(() => {
         this.host.project.refresh();
-      })
+      });
     }
   }
 });
